Move dsl/string/* to dsl/helpers/*.
