+++ lib/vedeu/dsl/helpers/text.rb
