+++ lib/vedeu/dsl/helpers/truncate.rb
