+++ lib/vedeu/dsl/helpers/wordwrap.rb
