+++ test/lib/vedeu/dsl/helpers/align_test.rb
