+++ test/lib/vedeu/dsl/helpers/text_test.rb
