+++ test/lib/vedeu/dsl/helpers/truncate_test.rb
